requirements.txt
albumentations==1.3.1
contextlib2==21.6.0
joblib==1.3.2
multiprocess==0.70.15
numba==0.58.1
numpy==1.26.1
nvgpu==0.10.0
nvidia-ml-py==12.535.108
opencv-python==4.8.1.78
packaging==21.3
pandas==1.5.3
pathos==0.3.1
Pillow==10.1.0
py4j==0.10.9.5
pyarrow==13.0.0
pyfunctional==1.4.3
PyYAML
safetensors==0.4.0
scikit-image==0.22.0
scikit-learn==1.3.2
scipy
torch==2.0.1
torchvision==0.15.2
tqdm 
ujson==5.8.0
Werkzeug==3.0.1
matplotlib==3.8.0
huggingface_hub==0.26.2
